requirements.txt
#--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-functions==1.17.0
openai==0.27.8
scipy==1.10.0
transformers==4.30.0
python-dotenv==1.0.0
azure-storage-blob==12.14.1
azure-identity==1.12.0
requests==2.31.0
tiktoken==0.2.0
#langchain==0.0.329
langchain==0.1.11
beautifulsoup4==4.12.0
fake-useragent==1.1.3
chardet==5.1.0
#azure-search-documents==11.4.0b8
azure-search-documents==11.4.0
opencensus-ext-azure==1.1.9
pandas==1.5.1
python-docx==0.8.11
python-jose==3.3.0
aiohttp==3.9.3
msal==1.28.0
tenacity==8.2.3
python-certifi-win32==1.6.1
psycopg==3.1.18
psycopg-binary==3.1.18
pytest==8.2.0
mock==5.1.0
pytest-cov==5.0.0
pytest-mock==3.14.0
azure-eventhub==5.11.5
azure-ai-formrecognizer==3.2.0
pytest-asyncio == 0.23.7
